third_party/blink/web_tests/external/wpt/svg/types/scripted/SVGGraphicsElement-padding-expected.txt

This is a testharness.js-based test.
[FAIL] getScreenCTM with padding-left, scale
  assert_equals: expected 47 but got 41
Harness: the test ran to completion.
